requirements.txt
pyacoustid==1.2.2